+++ apps/mis-server/src/app.ts
@@ -19,7 +19,7 @@ import { accountServiceServer } from "src/services/account";
 import { adminServiceServer } from "src/services/admin";
 import { chargingServiceServer } from "src/services/charging";
 import { configServiceServer } from "src/services/config";
-import { fileServiceServer } from "src/services/file";
+import { exportServiceServer } from "src/services/export";
 import { initServiceServer } from "src/services/init";
 import { jobServiceServer } from "src/services/job";
 import { jobChargeLimitServer } from "src/services/jobChargeLimit";
@@ -52,7 +52,7 @@ export async function createServer() {
   await server.register(tenantServiceServer);
   await server.register(configServiceServer);
   await server.register(misConfigServiceServer);
-  await server.register(fileServiceServer);
+  await server.register(exportServiceServer);
 
   await server.ext.syncBlockStatus.sync();
