--- protos/server/file.proto
@@ -1,55 +1,0 @@
-/**
- * Copyright (c) 2022 Peking University and Peking University Institute for Computing and Digital Economy
- * SCOW is licensed under Mulan PSL v2.
- * You can use this software according to the terms and conditions of the Mulan PSL v2.
- * You may obtain a copy of Mulan PSL v2 at:
- *          http://license.coscl.org.cn/MulanPSL2
- * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
- * EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
- * MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
- * See the Mulan PSL v2 for more details.
- */
-
-syntax = "proto3";
-
-package scow.server;
-
-import "server/charging.proto";
-import "server/user.proto";
-
-message ExportAccount {
-  optional string tenant_name = 1;
-  optional string account_name = 2;
-  // 为true时表明账户被冻结，false则为所有账户
-  optional bool blocked = 3;
-  // 为true时表明账户欠费，false则为所有账户
-  optional bool debt = 4;
-}
-
-message ExportUser {
-  optional string tenant_name = 1;
-  optional GetAllUsersRequest.UsersSortField sort_field = 3;
-  optional SortDirection sort_order = 4;
-  optional string id_or_name = 5;
-  optional PlatformRole platform_role = 6;
-  optional TenantRole tenant_role = 7;
-}
-
-message ExportRequest {
-  uint32 count = 1;
-  oneof export_event {
-    ExportUser user = 2;
-    ExportAccount account = 3;
-    GetChargeRecordsTotalCountRequest charge_record = 4;
-    GetPaymentRecordsRequest pay_record = 5;
-  }
-}
-
-message ExportResponse {
-  bytes chunk = 1;
-}
-
-
-service FileService {
-  rpc Export(ExportRequest) returns (stream ExportResponse);
-}
